+++ packages/react/components/badge.d.ts
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { Tooltip as TooltipNamespace } from 'framework7/components/tooltip/tooltip';
 
 declare namespace F7Badge {
   interface Props {
@@ -6,6 +7,8 @@ declare namespace F7Badge {
     id? : string | number
     className? : string
     style? : React.CSSProperties
+    tooltip? : string
+    tooltipTrigger? : string
     color? : string
     colorTheme? : string
     textColor? : string
@@ -17,6 +20,6 @@ declare namespace F7Badge {
   }
 }
 declare class F7Badge extends React.Component<F7Badge.Props, {}> {
-  
+  f7Tooltip: TooltipNamespace.Tooltip
 }
 export default F7Badge;