+++ packages/vue/components/badge.js
@@ -4,7 +4,9 @@ import __vueComponentProps from '../runtime-helpers/vue-component-props.js';
 export default {
   name: 'f7-badge',
   props: Object.assign({
-    id: [String, Number]
+    id: [String, Number],
+    tooltip: String,
+    tooltipTrigger: String
   }, Mixins.colorProps),
 
   render() {
@@ -17,12 +19,66 @@ export default {
     } = props;
     const classes = Utils.classNames(className, 'badge', Mixins.colorClasses(props));
     return _h('span', {
+      ref: 'el',
       style: style,
       class: classes,
       attrs: {
         id: id
       }
     }, [this.$slots['default']]);
+  },
+
+  watch: {
+    'props.tooltip': function watchTooltip(newText) {
+      const self = this;
+
+      if (!newText && self.f7Tooltip) {
+        self.f7Tooltip.destroy();
+        self.f7Tooltip = null;
+        delete self.f7Tooltip;
+        return;
+      }
+
+      if (newText && !self.f7Tooltip && self.$f7) {
+        self.f7Tooltip = self.$f7.tooltip.create({
+          targetEl: self.$refs.el,
+          text: newText,
+          trigger: self.props.tooltipTrigger
+        });
+        return;
+      }
+
+      if (!newText || !self.f7Tooltip) return;
+      self.f7Tooltip.setText(newText);
+    }
+  },
+
+  mounted() {
+    const self = this;
+    const el = self.$refs.el;
+    if (!el) return;
+    const {
+      tooltip,
+      tooltipTrigger
+    } = self.props;
+    if (!tooltip) return;
+    self.$f7ready(f7 => {
+      self.f7Tooltip = f7.tooltip.create({
+        targetEl: el,
+        text: tooltip,
+        trigger: tooltipTrigger
+      });
+    });
+  },
+
+  beforeDestroy() {
+    const self = this;
+
+    if (self.f7Tooltip && self.f7Tooltip.destroy) {
+      self.f7Tooltip.destroy();
+      self.f7Tooltip = null;
+      delete self.f7Tooltip;
+    }
   },
 
   computed: {