+++ packages/vue/components/chip.js
@@ -12,7 +12,9 @@ export default {
     deleteable: Boolean,
     mediaBgColor: String,
     mediaTextColor: String,
-    outline: Boolean
+    outline: Boolean,
+    tooltip: String,
+    tooltipTrigger: String
   }, Mixins.colorProps, {}, Mixins.linkIconProps),
 
   render() {
@@ -91,23 +93,70 @@ export default {
     }, [mediaEl, labelEl, deleteEl]);
   },
 
+  watch: {
+    'props.tooltip': function watchTooltip(newText) {
+      const self = this;
+
+      if (!newText && self.f7Tooltip) {
+        self.f7Tooltip.destroy();
+        self.f7Tooltip = null;
+        delete self.f7Tooltip;
+        return;
+      }
+
+      if (newText && !self.f7Tooltip && self.$f7) {
+        self.f7Tooltip = self.$f7.tooltip.create({
+          targetEl: self.$refs.el,
+          text: newText,
+          trigger: self.props.tooltipTrigger
+        });
+        return;
+      }
+
+      if (!newText || !self.f7Tooltip) return;
+      self.f7Tooltip.setText(newText);
+    }
+  },
+
   created() {
     Utils.bindMethods(this, ['onClick', 'onDeleteClick']);
   },
 
   mounted() {
-    this.$refs.el.addEventListener('click', this.onClick);
+    const self = this;
+    const el = self.$refs.el;
+    el.addEventListener('click', self.onClick);
 
-    if (this.$refs.deleteEl) {
-      this.$refs.deleteEl.addEventListener('click', this.onDeleteClick);
+    if (self.$refs.deleteEl) {
+      self.$refs.deleteEl.addEventListener('click', self.onDeleteClick);
     }
+
+    const {
+      tooltip,
+      tooltipTrigger
+    } = self.props;
+    if (!tooltip) return;
+    self.$f7ready(f7 => {
+      self.f7Tooltip = f7.tooltip.create({
+        targetEl: el,
+        text: tooltip,
+        trigger: tooltipTrigger
+      });
+    });
   },
 
   beforeDestroy() {
-    this.$refs.el.removeEventListener('click', this.onClick);
+    const self = this;
+    self.$refs.el.removeEventListener('click', self.onClick);
 
-    if (this.$refs.deleteEl) {
-      this.$refs.deleteEl.removeEventListener('click', this.onDeleteClick);
+    if (self.$refs.deleteEl) {
+      self.$refs.deleteEl.removeEventListener('click', self.onDeleteClick);
+    }
+
+    if (self.f7Tooltip && self.f7Tooltip.destroy) {
+      self.f7Tooltip.destroy();
+      self.f7Tooltip = null;
+      delete self.f7Tooltip;
     }
   },